// StaffSync/timemanagementsystem.client/src/enums/Roles.ts
enum Roles {
  HRManagerAdmin = "HR Manager Admin",
  Employee = "Employee",
  HRMSAdmin = "HRMS Admin",
  SuperAdmin = "Super Admin",
  SrOperationsManager = "Sr. Operations Manager",
  OPSSrTeamLeader = "OPS Sr. Team Leader",
  OPSTeamLeader = "OPS Team Leader",
  Manager = "Manager",
  SrManager = "Sr. Manager",
  SrSupervisor = "Sr. Supervisor",
  Supervisor = "Supervisor",
  ItManagerAdmin = "It Manager Admin",
}

export default Roles;